+++ changelog.md
@@ -1,1 +1,2 @@
-add Argentine Spanish (es_ar) translation, thanks to Texaliuz!
+- add option to render shadow below chat heads (on by default)
+- render upside down players upside down